+++ src/main/java/com/kyonggi/diet/dietFile/CSVController.java
@@ -1,6 +1,5 @@
 package com.kyonggi.diet.dietFile;
 
-import com.kyonggi.diet.aws.S3FileRequest;
 import com.kyonggi.diet.controllerDocs.CSVControllerDocs;
 import com.opencsv.exceptions.CsvValidationException;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -15,22 +14,13 @@ import java.io.IOException;
 @Tag(name = "CSV 파일 API", description = "CSV 파일을 읽기 위한 API 입니다.")
 public class CSVController implements CSVControllerDocs {
 
-    private final CSVService csvService;
-
-    private final com.kyonggi.diet.aws.CSVServices csvReader;
+    private final CSVService csvReader;
 
     @PostMapping("/diet")
-    public void readDiet(@RequestBody String path) throws CsvValidationException, IOException {
-        if(path == null)
+    public void readDiet(@RequestBody String key) throws CsvValidationException, IOException {
+        if(key == null)
             return;
-        csvService.readAndSave(path);
-    }
-
-    @PostMapping("/diet2")
-    public void readDiet2(@RequestBody S3FileRequest s3FileRequest) throws CsvValidationException, IOException {
-        if(s3FileRequest.getKey() == null)
-            return;
-        csvReader.readAndSave(s3FileRequest.getKey());
+        csvReader.readAndSave(key);
     }
 
 }
